Added damageable functionality to the bulbatoes so that they hurt the wallow boss.
If I want the bulbatoes to give the player health on kill I need to make them inherit from the enemy class and their damagable inherit from the enemydamagable class.
Need to add an on death effect for the boss, a post death screen, a title screen, a control scheme choice screen, need to refactor the level design, and update some QoL changes for the core gameplay (have the attacks work if they start within an enemy address a bug that if darwin is killed while moving his corpse moves. )