Added an end Credits and a restart RestartLevel button.
Also modified the controls file to allow for both keyboard and controller to work at the same time but it is buggy and needs another look.
NEED TO: FIX CONTROLLER BUG (If it is set to the opposite control scheme in the gui  the controls will be slow for that setting)
Make it so when the controller moves it highlights the restart button and a press restarts it, Currently only works with a mouse.
All of the small adjustments to the game mechanics.
Reenable joy demons
Redesign the level.
Add the hand power up.